pyproject.toml
[project]
name = "stimpack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "jedi-language-server>=0.46.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pyprojroot>=0.3.0",
    "pyzmq>=27.1.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]
[tool.setuptools]
packages = ["runtime"]
